# Conflicts:
#	app/Services/ContentManagementServices/CmsLanguageService.php

+++ database/migrations/2021_11_15_093805_create_static_page_contents_table.php
@@ -15,7 +15,7 @@ class CreateStaticPageContentsTable extends Migration
     {
         Schema::create('static_page_contents', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('static_page_type_id');
+            $table->unsignedSmallInteger('static_page_type_id');
             $table->unsignedTinyInteger('show_in')
                 ->comment('1=>Nise3, 2=> Youth, 3=>TSP, 4=>Industry, 5=>Industry Association');
 
@@ -35,7 +35,7 @@ class CreateStaticPageContentsTable extends Migration
             $table->unsignedInteger('created_by')->nullable();
             $table->unsignedInteger('updated_by')->nullable();
 
-            $table->tinyInteger('row_status')
+            $table->unsignedTinyInteger('row_status')
                 ->default(1)
                 ->comment('ACTIVE_STATUS => 1, INACTIVE_STATUS => 0');
